backs/requirements.txt
flask
flask-cors
pymongo
dnspython
python-dotenv
flask-bcrypt
pyjwt
